+++ selfdrive/controls/lib/drive_helpers.py
@@ -1,7 +1,6 @@
 from cereal import log
-from openpilot.common.numpy_fast import clip, interp
-from openpilot.common.realtime import DT_MDL, DT_CTRL
-from openpilot.sunnypilot.modeld.constants import ModelConstants
+from openpilot.common.numpy_fast import clip
+from openpilot.common.realtime import DT_CTRL
 
 MIN_SPEED = 1.0
 CONTROL_N = 17
@@ -20,29 +19,6 @@ def clip_curvature(v_ego, prev_curvature, new_curvature):
   safe_desired_curvature = clip(new_curvature,
                                 prev_curvature - max_curvature_rate * DT_CTRL,
                                 prev_curvature + max_curvature_rate * DT_CTRL)
-
-  return safe_desired_curvature
-
-
-def get_lag_adjusted_curvature(steer_delay, v_ego, psis, curvatures):
-  if len(psis) != CONTROL_N:
-    psis = [0.0]*CONTROL_N
-    curvatures = [0.0]*CONTROL_N
-  v_ego = max(MIN_SPEED, v_ego)
-
-  # MPC can plan to turn the wheel and turn back before t_delay. This means
-  # in high delay cases some corrections never even get commanded. So just use
-  # psi to calculate a simple linearization of desired curvature
-  current_curvature_desired = curvatures[0]
-  psi = interp(steer_delay, ModelConstants.T_IDXS[:CONTROL_N], psis)
-  average_curvature_desired = psi / (v_ego * steer_delay)
-  desired_curvature = 2 * average_curvature_desired - current_curvature_desired
-
-  # This is the "desired rate of the setpoint" not an actual desired rate
-  max_curvature_rate = MAX_LATERAL_JERK / (v_ego**2) # inexact calculation, check https://github.com/commaai/openpilot/pull/24755
-  safe_desired_curvature = clip(desired_curvature,
-                                current_curvature_desired - max_curvature_rate * DT_MDL,
-                                current_curvature_desired + max_curvature_rate * DT_MDL)
 
   return safe_desired_curvature
